fix: 파일 구조 개편
