+++ src/middleware/ImageMiddleware.js
