+++ src/middleware/ImageResizeMiddleware.js
